+++ src/java1/bonus00/bonus01.java
@@ -1,0 +1,11 @@
+package src.java1.bonus00;
+
+public class bonus01 {
+    public static int findLSL(String str){
+        return 0;
+    }
+
+    public static void main(String[] args) {
+
+    }
+}
