+++ src/java1/ex00/hw_1.java
@@ -1,0 +1,7 @@
+package src.java1.ex00;
+
+public class hw_1 {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
