+++ src/java1/ex01/hw_3.java
@@ -1,0 +1,9 @@
+package src.java1.ex01;
+
+public class hw_3 {
+    public static void main(String[] args) {
+        String name;
+        name  = "John";
+        System.out.println("Hello, " + name + "!");
+    }
+}
