+++ src/java1/ex02/hw_2.java
@@ -1,0 +1,12 @@
+package src.java1.ex02;
+
+public class hw_2 {
+    public static void main(String[] args) {
+        int a = 5;
+        float b = 3.14f;
+        String c = "str";
+        boolean d = true;
+
+        System.out.println("a = " + a + ", b = " + b + ", c = " +c + ", d = " +d);
+    }
+}
