+++ src/java1/ex03/hw_2.java
@@ -1,0 +1,15 @@
+package src.java1.ex03;
+
+public class hw_2 {
+
+    public static void main(String[] args) {
+        int a = 5;
+        int b = 10;
+
+        System.out.println("a+b = " + add(a,b));
+    }
+
+    public static int add(int a, int b){
+        return a+b;
+    }
+}
