+++ src/java1/ex03/hw_3.java
@@ -1,0 +1,12 @@
+package src.java1.ex03;
+
+public class hw_3 {
+    public static void main(String[] args) {
+        double radius = 10;
+        System.out.println("calcArea(radius) = " + calcArea(radius));
+    }
+
+    public static double calcArea(double r){
+        return r*r*3.14;
+    }
+}
