+++ src/java1/ex05/hw_1.java
@@ -1,0 +1,10 @@
+package src.java1.ex05;
+
+public class hw_1 {
+    public static void main(String[] args) {
+        int age = 0;
+
+        if(age > 20)
+            System.out.println("성인입니다.");
+    }
+}
