+++ src/java1/ex05/hw_3.java
@@ -1,0 +1,16 @@
+package src.java1.ex05;
+
+public class hw_3 {
+    public static void main(String[] args) {
+        int score;
+        score = 80;
+        if(score >= 90)
+            System.out.println("A");
+        else if(score >= 80)
+            System.out.println("B");
+        else if(score >= 70)
+            System.out.println("C");
+        else
+            System.out.println("D");
+    }
+}
