+++ src/java1/ex06/hw_2.java
@@ -1,0 +1,13 @@
+package src.java1.ex06;
+
+public class hw_2 {
+    public static void main(String[] args) {
+        for(int i =1; i <= 10; i++)
+        {
+            if(i%2 == 0)
+                System.out.println(i + " ");
+            else
+                continue;
+        }
+    }
+}
