+++ src/java1/ex06/hw_4.java
@@ -1,0 +1,16 @@
+package src.java1.ex06;
+
+public class hw_4 {
+    public static void main(String[] args) {
+        int idx = 1;
+        while(true)
+        {
+            if(idx > 100)
+                break;
+
+            if((idx%3==0)||(idx%4==0))
+                System.out.println(idx);
+            idx++;
+        }
+    }
+}
