--- src/java1/ex08/ex00/hw_1.java
@@ -1,7 +1,0 @@
-package ex00;
-
-public class hw_1 {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
