--- src/java1/ex08/ex02/hw_3.java
@@ -1,9 +1,0 @@
-package ex02;
-
-public class hw_3 {
-    public static void main(String[] args) {
-        //short s1 = 1, short s2 = 2, short s3 = s1+s2;
-        //한 줄에 여러개의 변수를 선언할수 없으며, s3는 s1+s2값을 short형으로 형변환 해줘야함
-        // 세미콜론 없음
-    }
-}
