--- src/java1/ex08/ex04/hw_2.java
@@ -1,12 +1,0 @@
-package ex04;
-
-public class hw_2 {
-    public static void main(String[] args) {
-        int x;
-        int y;
-        x = 5;
-        y = 3;
-
-        System.out.println("x+y = " + (x+y) + ", x-y = " + (x-y));
-    }
-}
