--- src/java1/ex08/ex04/hw_3.java
@@ -1,13 +1,0 @@
-package ex04;
-
-public class hw_3 {
-    public static void main(String[] args) {
-        int num1=15;
-        int num2=8;
-
-        if((num1%3 == 0)&&(num2%4==0))
-            System.out.println("True");
-        else
-            System.out.println("False");
-    }
-}
