--- src/java1/ex08/ex04/hw_4.java
@@ -1,11 +1,0 @@
-package ex04;
-
-public class hw_4 {
-    public static void main(String[] args) {
-        int x=5;
-        int y=3;
-
-        System.out.println("x * y = " + (x*y) + ", x / y = " + (x/y));
-
-    }
-}
