--- src/java1/ex08/ex04/hw_5.java
@@ -1,13 +1,0 @@
-package ex04;
-
-public class hw_5 {
-    public static void main(String[] args) {
-        int age;
-
-        age = 25;
-
-        if(age > 18)
-            System.out.println("성인입니다.");
-
-    }
-}
