--- src/java1/ex08/ex04/hw_6.java
@@ -1,10 +1,0 @@
-package ex04;
-
-public class hw_6 {
-    public static void main(String[] args) {
-        int[] arr = {1,2,3,4,5};
-
-        for(int i =0; i < arr.length; i++)
-            System.out.printf("%d ", arr[i]);
-    }
-}
