--- src/java1/ex08/ex05/hw_1.java
@@ -1,10 +1,0 @@
-package ex05;
-
-public class hw_1 {
-    public static void main(String[] args) {
-        int age = 0;
-
-        if(age > 20)
-            System.out.println("성인입니다.");
-    }
-}
