--- src/java1/ex08/ex05/hw_2.java
@@ -1,14 +1,0 @@
-package ex05;
-
-public class hw_2 {
-    public static void main(String[] args) {
-        int num = 15;
-
-        if(num % 2 == 0)
-            System.out.println("양수입니다.");
-        else if(num == 0)
-            System.out.println("0입니다.");
-        else
-            System.out.println("음수입니다.");
-    }
-}
