--- src/java1/ex08/ex05/hw_5.java
@@ -1,10 +1,0 @@
-package ex05;
-
-public class hw_5 {
-    public static void main(String[] args) {
-        int num1=10;
-        int num2=5;
-
-        System.out.println(num1>num2? num1:num2);
-    }
-}
