--- src/java1/ex08/ex06/hw_2.java
@@ -1,13 +1,0 @@
-package ex06;
-
-public class hw_2 {
-    public static void main(String[] args) {
-        for(int i =1; i <= 10; i++)
-        {
-            if(i%2 == 0)
-                System.out.println(i + " ");
-            else
-                continue;
-        }
-    }
-}
