--- src/java1/ex08/ex06/hw_3.java
@@ -1,13 +1,0 @@
-package ex06;
-
-public class hw_3 {
-    public static void main(String[] args) {
-        for(int i = 1; i < 10; i++)
-        {
-            for(int j = 1; j < 10; j++)
-            {
-                System.out.println(i + " x " + j + " = " + (i*j));
-            }
-        }
-    }
-}
