+++ src/java1/ex08/hw_3.java
@@ -1,4 +1,4 @@
-package ex08;
+package src.java1.ex08;
 
 import java.util.Scanner;
